Home

PRISMA.PRISMA — Module

PRISMA.jl

Julia package for generating checklists and flow diagrams based on the 2020 Preferred Reporting Items for Systematic Reviews and Meta-Analyses (PRISMA) statement (Page et al., 2021).

New Functions and Types:

For working with PRISMA checklists:

checklist_dataframe - returns an empty PRISMA checklist as the type DataFrame

checklist_template - returns an empty PRISMA checklist as comma-separated values

checklist_read - reads the checklist data as a DataFrame from comma-separated values

checklist - takes a PDF and returns a completed PRISMA checklist as the type Checklist

checklist_save - saves a Checklist as comma-separated values

Checklist - includes a completed checklist and paper metadata

For working with PRISMA flow diagrams:

flow_diagram_dataframe - returns the template PRISMA flow diagram data as a DataFrame

flow_diagram_template - returns the template PRISMA flow diagram data as comma-separated values

flow_diagram_read - reads the flow diagram data as a DataFrame from comma-separated values

flow_diagram - returns a PRISMA flow diagram as the type FlowDiagram

flow_diagram_save - saves a FlowDiagram using Graphviz

FlowDiagram - the flow diagram type that can be displayed in the plot panel or saved

Reexports

OrderedCollections.OrderedDict - Checklist.metadata returns an OrderedDict

DataFrames.DataFrame - both flow_diagram_dataframe and checklist_dataframe return a DataFrame

Base.display - PRISMA adds a method for displaying the FlowDiagram in the plot panel

Base.show - PRISMA adds new methods for printing the Checklist and FlowDiagram types