Checklist

PRISMA.checklist_df — Function

PRISMA.checklist_df()::DataFrame

Returns a template PRISMA checklist as a DataFrame

Examples

You can export the DataFrame to an Excel spreadsheet or CSV file and edit the checklist using a spreadsheet program. You can also edit the DataFrame directly in a Julia program using the DataFrames.jl package.

using PRISMA

df = checklist_df()

checklist_save("checklist.csv",  df)
checklist_save("checklist.xlsx", df)
checklist_save("checklist.html", df; editable=true)
checklist_save("checklist.json", df)

using PRISMA

df = checklist_df()
df[3, "Location where item is reported"] = "Sysemtatic review is in the title."
df[3, "Location where item is reported"] = "The completed abastract is located on page one."

Additionally, the checklist() function uses checklist_df() as a template but takes a paper in PDF format, parses it using natural language processing via Flux.jl and Transformers.jl, and returns a completed checklist along with the paper's metadata, represented by the type Checklist.

PRISMA.checklist — Function

PRISMA.checklist(paper::AbstractString)::Checklist
PRISMA.checklist(bytes::Vector{UInt8})::Checklist

Returns a completed PRISMA checklist as the type Checklist. The Checklist type includes a completed checklist as a DataFrame and the metadata of the paper as a LittleDict. The paper argument can be a path to a pdf file or an array of bytes. This function uses the C++ library Poppler via Poppler_jll to parse the pdf and the natural language processing functionality in Julia via Flux.jl and Transformers.jl to find items from the checklist in the paper and populate the Comments or location in manuscript and Yes/No/NA columns in the DataFrame from checklist_df().

The following metadata is parsed from the pdf file and stored in the LittleDict as:

"title": the title of the paper

"subject": the subject of the paper

"author": the author of the paper

"creator": the creator of the paper

"producer": the producer of the paper

"creation date": the date the paper was created

"modification date": the date the paper was last modified

"pages": the number of pages in the paper

"paper size": the size of the paper

"paper rotation": the rotation of the paper

"file size": the size of the pdf file

"optimized": whether the pdf was optimized

"pdf version": the version of the pdf

All keys and values in the dictionary ar eof type String. If the parsing fails the value will be an empty string.

Arguments

paper::AbstractString: a path to a pdf file as a string

bytes::Vector{UInt8}: the pdf data as an array of bytes

Returns

Checklist: a completed checklist with paper metadata

Examples

Export a completed checklist to an Microsoft® Excel spreadsheet. Just make the sure the file path end with .xlsx.

using PRISMA

clist = checklist("manuscript.pdf")

checklist_save("checklist.xlsx", clist)

It is also possible to view the completed checklist directly in a Julia program and using the DataFrames.jl package to edit the DataFrame directly to add values that the fubction could not find or to add new items or quick edits.

using PRISMA, DataFrames

df::DataFrame = checklist("manuscript.pdf").df

println(df)

df[02, "Yes/No/NA"] = "Yes"
df[09, "Yes/No/NA"] = "Yes"
df[10, "Yes/No/NA"] = "Yes"

PRISMA.checklist_read — Function

checklist_read(fn::AbstractString; sheetname::AbstractString="2020 PRISMA Checklist", kwargs....)::DataFrame

reads the template data from a CSV, XLSX, HTML, or JSON

Arguments

fn::AbstractString: the name of the file to read

sheetname::AbstractString="2020 PRISMA Checklist": the name of the sheet in the spreadsheet

kwargs...: additional arguments to be passed to the underlying

CSV.read, XLSX.readtable, HTMLTables.read, and JSON3.read functions

Returns

DataFrame: the template dataframe

Examples

using PRISMA: checklist_save, checklist_df, checklist_read

checklist_save("checklist.csv",  checklist_df())
checklist_save("checklist.xlsx", checklist_df())
checklist_save("checklist.json", checklist_df())
checklist_save("checklist.html", checklist_df())

checklist_read("checklist.csv")
checklist_read("checklist.xlsx")
checklist_read("checklist.json")
checklist_read("checklist.html")

PRISMA.Checklist — Type

PRISMA.Checklist

This types represents a PRISMA checklist in the form of a DataFrame and the metadata of the paper that was used to generate it as a LittleDict or an small ordered dictionary withs keys and values.

Keyword Arguments

df::DataFrame: the checklist as a DataFrame

metadata::LittleDict: the metadata of the paper

Examples

using PRISMA

cl::Checklist = checklist("paper.pdf")

title = cl.metadata["title"]
println(title)

pages = cl.metadata["pages"]
println(pages)

PRISMA.checklist_save — Function

checklist_save(
    fn::AbstractString, 
    cl::Checklist; 
    sheetname::AbstractString="2020 PRISMA Checklist", 
    overwrite::Bool=false, 
    kwargs...
)::String

saves as either Checklist a JSON, CSV, XLSX, or HTML.

Arguments

fn::AbstractString: the name of the file to save

cl::Checklist: the checklist to save

Keyword Arguments

sheetname::AbstractString="2020 PRISMA Checklist": the name of the sheet in the spreadsheet

overwrite::Bool: whether to overwrite the file if it already exists

kwargs...: arguments to be passed to the underlying CSV.write, XLSX.writetable,

HTMLTables.write, and JSON3.write functions

Returns

String: the path to the saved file

Examples

using PRISMA: Checklist, checklist, checklist_save

clist::Checklist = checklist("manuscript.pdf")

checklist_save("checklist.csv",  clist)
checklist_save("checklist.xlsx", clist)
checklist_save("checklist.json", clist)
checklist_save("checklist.html", clist)

checklist_save(
    fn::AbstractString, 
    cl::Checklist; 
    sheetname::AbstractString="2020 PRISMA Checklist", 
    overwrite::Bool=false, 
    kwargs...
)::String

saves as either Checklist a JSON, CSV, XLSX, or HTML.

Arguments

fn::AbstractString: the name of the file to save

df::DataFrame=checklist_df(): the dataframe to save

Keyword Arguments

sheetname::AbstractString="2020 PRISMA Checklist": the name of the sheet in the spreadsheet

overwrite::Bool: whether to overwrite the file if it already exists

kwargs...: arguments to be passed to the underlying CSV.write, XLSX.writetable,

HTMLTables.write, and JSON3.write functions

Returns

String: the path to the saved file

Examples

using PRISMA: checklist, checklist_save
using DataFrames: DataFrame

df::DataFrame = checklist_df()

checklist_save("checklist.csv",  df)
checklist_save("checklist.xlsx", df)
checklist_save("checklist.json", df)
checklist_save("checklist.html", df)