Checklist

PRISMA.checklist_df — Function

PRISMA.checklist_df()::DataFrame

Returns a template PRISMA checklist as a DataFrame

You can export the DataFrame to an Excel spreadsheet or CSV file and edit the checklist using a spreadsheet program. You can also edit the DataFrame directly in a Julia program using the DataFrames.jl package.

Examples

using PRISMA, CSV

CSV.write("checklist.csv", checklist_df())

using PRISMA, XLSX

XLSX.writetable("checklist.xlsx", "PRISMA Checklist" => checklist_df())

using PRISMA, DataFrames

df = checklist_df()
df[3, "Location where item is reported"] = "Sysemtatic review is in the title."
df[3, "Location where item is reported"] = "The completed abastract is located on page one."

Additionally, the checklist() function uses checklist_df() as a template but takes a paper in PDF format, parses it using natural language processing via Flux.jl and Transformers.jl, and returns a completed checklist along with the paper's metadata, represented by the type Checklist.

PRISMA.checklist — Function

PRISMA.checklist(paper::AbstractString)::Checklist
PRISMA.checklist(bytes::Vector{UInt8})::Checklist

Returns a completed PRISMA checklist as the type Checklist. The Checklist type includes a completed checklist as a DataFrame and the metadata of the paper as a LittleDict. The paper argument can be a path to a pdf file or an array of bytes. This function uses the C++ library Poppler via Poppler_jll to parse the pdf and the natural language processing functionality in Julia via Flux.jl and Transformers.jl to find items from the checklist in the paper and populate the Comments or location in manuscript and Yes/No/NA columns in the DataFrame from checklist_df().

The following metadata is parsed from the pdf file and stored in the LittleDict as:

"title": the title of the paper

"subject": the subject of the paper

"author": the author of the paper

"creator": the creator of the paper

"producer": the producer of the paper

"creation date": the date the paper was created

"modification date": the date the paper was last modified

"pages": the number of pages in the paper

"paper size": the size of the paper

"paper rotation": the rotation of the paper

"file size": the size of the pdf file

"optimized": whether the pdf was optimized

"pdf version": the version of the pdf

All keys and values in the dictionary ar eof type String. If the parsing fails the value will be an empty string.

Arguments

paper::AbstractString: a path to a pdf file as a string

bytes::Vector{UInt8}: the pdf data as an array of bytes

Returns

Checklist: a completed checklist with paper metadata

Examples

Export a completed checklist to an Microsoft® Excel spreadsheet using the XLSX package.

using PRISMA, XLSX

clist = checklist("manuscript.pdf")

XLSX.writetable("checklist.xlsx", "PRISMA Checklist" => clist.df)

Export a completed checklist as a CSV file using the CSV package.

using PRISMA, CSV

clist = checklist("manuscript.pdf")

CSV.write("checklist.csv", clist.df)

It is also possible to view the completed checklist directly in a Julia program and using the DataFrames.jl package to edit the DataFrame directly to add values that the fubction could not find or to add new items or quick edits.

using PRISMA, DataFrames

clist = checklist("manuscript.pdf")

println(clist.df)

clist.df[02, "Yes/No/NA"] = "Yes"
clist.df[09, "Yes/No/NA"] = "Yes"
clist.df[10, "Yes/No/NA"] = "Yes"

PRISMA.Checklist — Type

PRISMA.Checklist

This types represents a PRISMA checklist in the form of a DataFrame and the metadata of the paper that was used to generate it as a LittleDict or an small ordered dictionary withs keys and values.

Keyword Arguments

df::DataFrame: the checklist as a DataFrame

metadata::LittleDict: the metadata of the paper