Flow Diagram

PRISMA.flow_diagram_df — Function

PRISMA.flow_diagram_df()::DataFrame

Return the template that is used to create the flow diagram as a DataFrame.

Examples

You can either edit the template in a Julia program or save it as a CSV file.

using PRISMA, CSV

df = PRISMA.flow_diagram_df()

CSV.write("flow_diagram.csv", df)

using PRISMA, DataFrames

df = PRISMA.flow_diagram_df()

df[3, "result"] = 200

In order to add more databases or registers to the flow diagram, you must add them as another row. This can either be done by saving the template as a CSV file and editing it in a Julia program using the DataFrames package.

When you want to add a specific text to the a specific box the box must be specificied in the box_num column. There are 22 boxes in the flow diagram. When you using the DataFrames package, you can add the new row to the DataFrame using the push! function.

using PRISMA, DataFrames

df = PRISMA.flow_diagram_df()

# adding new Databases
push!(df, (8, "Database or Register 4", 150))
push!(df, (8, "Database or Register 5", 121))

# adding new reasons for exclusion
push!(df, (15, "Reason 4", 20))
push!(df, (15, "Reason 5", 10))
push!(df, (15, "Reason 6", 03))

println(df)

PRISMA.flow_diagram(df)

When you want to plot the data in the flow diagram, you can use the PRISMA.flow_diagram function. If your data was saved as a CSV file, you can use must read it in using the DataFrames package.

using PRISMA, CSV, DataFrames

df = CSV.read("flow_diagram.csv", DataFrame)

PRISMA.flow_diagram(df)

PRISMA.flow_diagram — Function

PRISMA.flow_diagram(
    data::DataFrame=flow_diagram_df(),
    background_color::AbstractString="white",
    grayboxes::Bool=true,
    grayboxes_color::AbstractString="#f0f0f0",
    top_boxes::Bool=true,
    top_boxes_borders::Bool=false,
    top_boxes_color::AbstractString="#ffc000",
    side_boxes::Bool=true,
    side_boxes_borders::Bool=false,
    side_boxes_color::AbstractString="#95cbff",
    previous_studies::Bool=true,
    other_methods::Bool=true,
    borders::Bool=true,
    border_style::AbstractString="solid",
    border_width::Union{AbstractString,Number}=1,
    border_color::AbstractString="black",
    font::AbstractString="Helvetica",
    font_color::AbstractString="black",
    font_size::Union{AbstractString,Number}=1,
    arrow_head::AbstractString="normal",
    arrow_size::Union{AbstractString,Number}=1,
    arrow_color::AbstractString="black",
    arrow_width::Union{AbstractString,Number}=1)::PRISMA.FlowDiagram

Generates the flow diagram figure from the flow diagram dataframe.

Argument

data::DataFrame: The flow diagram dataframe. Default is flow_diagram_df().

Keyword Arguments

background_color::String: The background color of the flow diagram. Default is white.

grayboxes::Bool: Whether to show gray boxes. Default is true.

grayboxes_color::String: The color of the gray boxes. Default is #f0f0f0.

top_boxes::Bool: Whether to show top boxes. Default is true.

top_boxes_border::Bool: Whether to show top boxes border. Default is false.

top_boxes_color::String: The color of the top boxes. Default is #ffc000.

side_boxes::Bool: Whether to show side boxes. Default is true.

side_boxes_border::Bool: Whether to show side boxes border. Default is false.

side_boxes_color::String: The color of the side boxes. Default is #95cbff.

previous_studies::Bool: Whether to show previous studies. Default is true.

other_methods::Bool: Whether to show other methods. Default is `

box_border_width::Number: The border width of the boxes. Default is 1.

box_border_color::String: The border color of the boxes. Default is black.

font::String: The font of the text. Default is Helvetica.

font_color::String: The color of the text. Default is black.

font_size::Number: The font size of the text. Default is 10.

arrow_color::String: The color of the arrows. Default is black.

arrow_width::Number: The width of the arrows. Default is 1.

Returns

PRISMA.FlowDiagram: The flow diagram figure.

Example

using PRISMA, CSV, DataFrame

# create a template to edit the data in a csv
template_df::DataFrame = PRISMA.flow_diagram_df()
CSV.write("flow_diagram.csv", template_df)

# create a `DataFrame` from the csv
df::DataFrame = CSV.read("flow_diagram.csv", DataFrame)

# generate the flow diagram with the `DataFrame`
fd::PRISMA.FlowDiagram = PRISMA.flow_diagram(df, top_boxes_color="white")

# save the flow diagram
PRISMA.flow_diagram_save("flow_diagram.svg", fd)

PRISMA.FlowDiagram — Type

PRISMA.FlowDiagram

Flow diagram type for PRISMA.jl

Argument

dot::AbstractString: The flow diagram written in Graphviz's DOT language

PRISMA.flow_diagram_save — Function

PRISMA.flow_diagram_save(fn::AbstractString, fd::FlowDiagram; overwrite::Bool=false)::String

writes a FlowDiagram as either a file (i.e., any Graphviz supported format) or stdout

Arguments

fn::AbstractString: The name of the file to be saved.

fd::FlowDiagram: The flow diagram to be saved.

overwrite::Bool=false: Whether to overwrite the file if it already exists.

Returns

String: The path to the saved file.

Examples

using PRISMA

fd = PRISMA.flow_diagram()

PRISMA.flow_diagram_save("flow_diagram.pdf", fd)
PRISMA.flow_diagram_save("flow_diagram.png", fd)
PRISMA.flow_diagram_save("flow_diagram.svg", fd)